improve code

+++ src/controllers/playlist.controller.ts
@@ -1,8 +1,9 @@
-import { isValidObjectId } from "mongoose"
+import mongoose, { isValidObjectId } from "mongoose"
 import { ApiError } from "../utils/ApiError"
 import { ApiResponse } from "../utils/ApiResponse"
 import { asyncHandler } from "../utils/asyncHandler"
 import { Playlist } from "../models/playlist.models.js"
+import { CustomUser } from "../type"
 
 
 const createPlaylist = asyncHandler(async (req, res) => {
@@ -30,7 +31,84 @@ const getUserPlaylists = asyncHandler(async (req, res) => {
 
   if (!isValidObjectId(userId)) throw new ApiError(400, 'userId is invalid')
 
-  const playlists = await Playlist.find({ owner: userId })
+  const playlists = await Playlist.aggregate(
+    [
+      {
+        $match: {
+          owner: new mongoose.Types.ObjectId(userId)
+        }
+      },
+      {
+        $lookup: {
+          from: "videos",
+          localField: "videos",
+          foreignField: "_id",
+          as: "videos",
+          pipeline: [
+            {
+              $lookup: {
+                from: "users",
+                localField: "owner",
+                foreignField: "_id",
+                as: "owner"
+              }
+            },
+            {
+              $addFields: {
+                owner: {
+                  $first: "$owner"
+                }
+              }
+            },
+            {
+              $project: {
+                title: 1,
+                thumbnail: 1,
+                description: 1,
+                owner: {
+                  username: 1,
+                  fullName: 1,
+                  avatar: 1,
+                }
+              }
+            }
+          ]
+        }
+      },
+      {
+        $lookup: {
+          from: "users",
+          localField: "owner",
+          foreignField: "_id",
+          as: "createdBy",
+          pipeline: [
+            {
+              $project: {
+                avatar: 1,
+                fullName: 1,
+                username: 1
+              }
+            }
+          ]
+        }
+      },
+      {
+        $addFields: {
+          createdBy: {
+            $first: "$createdBy"
+          }
+        }
+      },
+      {
+        $project: {
+          videos: 1,
+          createdBy: 1,
+          name: 1,
+          description: 1
+        }
+      }
+    ]
+  )
 
   if (!playlists) throw new ApiError(400, 'Error while fetching playlists')
 
@@ -46,7 +124,87 @@ const getPlaylistById = asyncHandler(async (req, res) => {
   if (!playlistId) throw new ApiError(400, 'playlistId is required')
   if (!isValidObjectId(playlistId)) throw new ApiError(400, 'playlistId is invalid')
 
-  const playlist = await Playlist.findById(playlistId)
+  const playlist = await Playlist.aggregate([
+    {
+      $match: {
+        _id: new mongoose.Types.ObjectId(playlistId)
+      }
+    },
+    {
+      //owner
+      $lookup: {
+        from: "users",
+        localField: "owner",
+        foreignField: "_id",
+        as: "createdBy",
+        pipeline: [
+          {
+            $project: {
+              fullName: 1,
+              username: 1,
+              avatar: 1
+            }
+          }
+        ]
+      }
+    },
+    {
+      $addFields: {
+        createdBy: {
+          $first: "$createdBy"
+        }
+      }
+    },
+    {
+      //videos
+      $lookup: {
+        from: "videos",
+        foreignField: "_id",
+        localField: "videos",
+        as: "videos",
+        pipeline: [
+          {
+            $lookup: {
+              from: "users",
+              localField: "owner",
+              foreignField: "_id",
+              as: "owner"
+            }
+          },
+          {
+            $addFields: {
+              owner: {
+                $first: "$owner"
+              }
+            }
+          },
+          {
+            $project: {
+              thumbnail: 1,
+              title: 1,
+              duration: 1,
+              views: 1,
+              owner: {
+                fullName: 1,
+                username: 1,
+                avatar: 1
+              },
+              createdAt: 1,
+              updatedAt: 1
+            }
+          }
+        ]
+      }
+    },
+    {
+      $project: {
+        name: 1,
+        description: 1,
+        videos: 1,
+        createdBy: 1
+      }
+    }
+  ])
 
   if (!playlist) throw new ApiError(400, 'playlist not found')
 
@@ -56,7 +214,7 @@ const getPlaylistById = asyncHandler(async (req, res) => {
 
 })
 
-const addVideoToPlaylist = asyncHandler(async (req, res) => {
+const addVideoToPlaylist = asyncHandler(async (req: CustomUser, res) => {
   const { playlistId, videoId } = req.params
 
   if (!playlistId || !videoId) throw new ApiError(400, 'playlistId and videoId are required')
@@ -68,6 +226,10 @@ const addVideoToPlaylist = asyncHandler(async (req, res) => {
   if (!playlist) throw new ApiError(400, 'playlist not found')
 
   if (playlist.videos.includes(videoId)) throw new ApiError(400, 'video already in playlist')
+
+  // check user is authorized to add video to playlist
+
+  if (playlist.owner.toString() !== req.user._id.toString()) throw new ApiError(400, 'Unauthorized')
 
   const updatedPlaylist = await Playlist.findByIdAndUpdate(playlistId, {
     $push: { videos: videoId }
@@ -81,7 +243,7 @@ const addVideoToPlaylist = asyncHandler(async (req, res) => {
 
 })
 
-const removeVideoFromPlaylist = asyncHandler(async (req, res) => {
+const removeVideoFromPlaylist = asyncHandler(async (req: CustomUser, res) => {
   const { playlistId, videoId } = req.params
 
   if (!playlistId || !videoId) throw new ApiError(400, 'playlistId and videoId are required')
@@ -93,6 +255,8 @@ const removeVideoFromPlaylist = asyncHandler(async (req, res) => {
   if (!playlist) throw new ApiError(400, 'playlist not found')
 
   if (!playlist.videos.includes(videoId)) throw new ApiError(400, 'video not in playlist')
+
+  if (playlist.owner.toString() !== req.user._id.toString()) throw new ApiError(400, 'Unauthorized')
 
   const updatedPlaylist = await Playlist.findByIdAndUpdate(playlistId, {
     $pull: { videos: videoId }
@@ -106,7 +270,7 @@ const removeVideoFromPlaylist = asyncHandler(async (req, res) => {
 
 })
 
-const deletePlaylist = asyncHandler(async (req, res) => {
+const deletePlaylist = asyncHandler(async (req: CustomUser, res) => {
   const { playlistId } = req.params
 
   if (!playlistId) throw new ApiError(400, 'playlistId is required')
@@ -116,6 +280,8 @@ const deletePlaylist = asyncHandler(async (req, res) => {
   const playlist = await Playlist.findById(playlistId)
 
   if (!playlist) throw new ApiError(400, 'playlist not found')
+
+  if (playlist.owner.toString() !== req.user._id.toString()) throw new ApiError(400, 'Unauthorized')
 
   const deletedPlaylist = await Playlist.findByIdAndDelete(playlistId)
 
@@ -127,7 +293,7 @@ const deletePlaylist = asyncHandler(async (req, res) => {
 
 })
 
-const updatePlaylist = asyncHandler(async (req, res) => {
+const updatePlaylist = asyncHandler(async (req: CustomUser, res) => {
   const { playlistId } = req.params
   const { name, description } = req.body
 
@@ -138,6 +304,9 @@ const updatePlaylist = asyncHandler(async (req, res) => {
   const playlist = await Playlist.findById(playlistId)
 
   if (!playlist) throw new ApiError(400, 'playlist not found')
+
+  if (playlist.owner.toString() !== req.user._id.toString()) throw new ApiError(400, 'Unauthorized')
+
 
   const updatedPlaylist = await Playlist.findByIdAndUpdate(playlistId, {
     name,
